subtype: add fast-path required by new optimizer
This is a new subtype pre-process step that attempts to look for
easily computable reasons that the result must be false (or true).
Since it can avoid allocating an `stenv` and otherwise setting up the
state needed for handling TypeVar constraints precisely,
it is hoped to save time (on average) by filtering out easy cases.